% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bakshaev_rudzkis}
\alias{bakshaev_rudzkis}
\title{Run Bakshaev and Rudzkis Test}
\usage{
bakshaev_rudzkis(dta, rnull, p, m_eval = 100L, nsim = 200L, nsim_mc = 1000L)
}
\arguments{
\item{dta}{data matrix.}

\item{rnull}{generate new data.}

\item{p}{, parameters for parametric bootstrap.}

\item{m_eval}{=100, number of evaluation points of kde.}

\item{nsim}{=200, number of simulation runs.}

\item{nsim_mc}{=1000, number of simulation runs.}
}
\value{
a list
}
\description{
Run Bakshaev and Rudzkis Test
}
