% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcTS}
\alias{calcTS}
\title{This function calculates the test statistics for  data}
\usage{
calcTS(dta, TS, typeTS, TSextra)
}
\arguments{
\item{dta}{data set (a matrix)}

\item{TS}{routine}

\item{typeTS}{format of TS}

\item{TSextra}{list passed to TS function}
}
\value{
A vector of values of test statistic(s)
}
\description{
This function calculates the test statistics for  data
}
\keyword{internal}
