% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.studies.cont.D5.R
\name{case.studies.cont.D5}
\alias{case.studies.cont.D5}
\title{Create various case studies for continuous data in 5 dimensions without parameter estimation}
\usage{
case.studies.cont.D5(which, nsample = 250, ReturnCaseNames = FALSE)
}
\arguments{
\item{which}{name of the case study.}

\item{nsample}{=250, sample size.}

\item{ReturnCaseNames}{=FALSE, just return names of case studies?}
}
\value{
a list of functions
}
\description{
This function creates the functions needed to run the various case studies.
}
