% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.studies.disc.R
\name{case.studies.disc}
\alias{case.studies.disc}
\title{Discretize 2D data from case studies}
\usage{
case.studies.disc(
  which,
  WithEstimation = FALSE,
  bintype = "ES",
  nbins = c(5, 5),
  nsample = 250
)
}
\arguments{
\item{which}{name or number of desired case study.}

\item{WithEstimation}{= FALSE, case study with or without parameter estimation.}

\item{bintype}{= "ES", ES for equal spaced bins or EP for equal probability bins.}

\item{nbins}{=c(5, 5) number of bins to use in x and y direction}

\item{nsample}{= 250, required sample size}
}
\value{
a list with needed stuff
}
\description{
This function provides the info necessary to run the case studies for discrete data.
}
