% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.studies.est.R
\name{case.studies.est}
\alias{case.studies.est}
\title{Create various case studies with parameter estimation}
\usage{
case.studies.est(which, nsample = 250, ReturnCaseNames = FALSE)
}
\arguments{
\item{which}{name of the case study.}

\item{nsample}{=250, sample size.}

\item{ReturnCaseNames}{=FALSE, just return names of case studies?}
}
\value{
a list of functions
}
\description{
This function creates the functions needed to run the various case studies that include parameter estimation.
}
