% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.studies.R
\name{case.studies}
\alias{case.studies}
\title{Create various case studies}
\usage{
case.studies(
  which,
  Continuous = TRUE,
  WithEstimation = FALSE,
  Dim = 2,
  nsample = 250,
  nbins = c(5, 5),
  bintype = "ES",
  ReturnCaseNames = FALSE
)
}
\arguments{
\item{which}{name or number of the case study.}

\item{Continuous}{= TRUE for continuous data}

\item{WithEstimation}{=FALSE, with parameter estimation}

\item{Dim}{=2 dimension of data}

\item{nsample}{=250, sample size.}

\item{nbins}{= c(5,5), number of bins for discretized data}

\item{bintype}{= "ES", ES for equal spaced bins or EP for equal probability bins.}

\item{ReturnCaseNames}{=FALSE, just return names of case studies?}
}
\value{
a list of functions
}
\description{
This function creates the functions needed to run the various case studies.
}
