% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.marginals.R
\name{change.marginals}
\alias{change.marginals}
\title{Change Marginals of 2D Data}
\usage{
change.marginals(which, side, nsample = 250, null_param)
}
\arguments{
\item{which}{method for modifying the data sets.}

\item{side}{which dimension to modify}

\item{nsample}{=250 sample size}

\item{null_param}{parameters under null hypothesis}
}
\value{
a list of functions
}
\description{
This function creates routines to modify the marginals
}
\keyword{internal}
