% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.functions.R
\name{check.functions}
\alias{check.functions}
\title{Sanity Checks}
\usage{
check.functions(pnull, rnull, phat = function(x) -99, x)
}
\arguments{
\item{pnull}{cdf under the null hypothesis}

\item{rnull}{routine to generate data under the null hypothesis}

\item{phat}{=function(x) -99, function to estimate parameters from the data, or -99}

\item{x}{matrix with data}
}
\description{
This function checks whether the inputs have the correct format
}
