% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq2D_test.R
\name{chi_disc_test}
\alias{chi_disc_test}
\title{Chi-square test for discrete 2D data}
\usage{
chi_disc_test(
  dta,
  pnull,
  dnull,
  phat = function(x) -99,
  minexpcount = 5,
  SuppressMessages = FALSE
)
}
\arguments{
\item{dta}{a matrix of numbers.}

\item{pnull}{distribution function to calculate expected counts.}

\item{dnull}{density to calculate expected counts.}

\item{phat}{=function(x) -99, function to estimate parameters of pnull.}

\item{minexpcount}{=5 minimum counts required per bin}

\item{SuppressMessages}{=TRUE, should info be shown?}
}
\value{
a vector with statistic, p value and degree of freedom
}
\description{
This function does the chi square goodness-of-fit test for discrete data in two dimensions.
}
