% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_M_for_dtaset}
\alias{compute_M_for_dtaset}
\title{Compute M statistic for one dtaset}
\usage{
compute_M_for_dtaset(dta, Eval, hs, rnull, p, nsim_mc)
}
\arguments{
\item{dta}{data matrix.}

\item{Eval}{matrix of evaluations}

\item{hs}{bandwiths}

\item{rnull}{generate new data}

\item{p}{values for parametric bootstrap}

\item{nsim_mc}{number of simulation runs}
}
\value{
a double
}
\description{
Compute M statistic for one dtaset
}
