% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discretize}
\alias{discretize}
\title{Bins continuous data}
\usage{
discretize(x, Range, nbins, ChangeVals = FALSE)
}
\arguments{
\item{x}{A numeric matrix with two columns.}

\item{Range}{range of variables}

\item{nbins}{number of bins.}

\item{ChangeVals}{=FALSE, should values of discrete rv's be adjusted to midpoints?}
}
\value{
A numeric matrix
}
\description{
Bins continuous data
}
