% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_case.R
\name{draw_case}
\alias{draw_case}
\title{Create plot for any case study}
\usage{
draw_case(
  which,
  Continuous = TRUE,
  WithEstimation = FALSE,
  Dim = 2,
  palt,
  nsample = 1000,
  Dms = c(1, 2),
  AltOnly = FALSE
)
}
\arguments{
\item{which}{name or number of the case study.}

\item{Continuous}{= TRUE for continuous data}

\item{WithEstimation}{=FALSE, with parameter estimation}

\item{Dim}{=2 dimension of data}

\item{palt}{parameter for alternative. If missing value in study is used.}

\item{nsample}{=250, sample size.}

\item{Dms}{=c(1,2, which dimensions are to be shown (for 5D data).}

\item{AltOnly}{= FALSE show only graph for alternative?}
}
\description{
This function illustrates any of the case studies.
}
