% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimateEV}
\alias{estimateEV}
\title{Estimate E and Var/n at Eval for given h, using MC from rnull}
\usage{
estimateEV(rnull, p, Eval, h, nsim_mc, n)
}
\arguments{
\item{rnull}{generate data under the null hypothesis.}

\item{p}{values for rnull}

\item{Eval}{matrix of evaluations}

\item{h}{bandwith, a double}

\item{nsim_mc}{number of simulation runs}

\item{n}{sample size}
}
\value{
a matrix
}
\description{
Estimate E and Var/n at Eval for given h, using MC from rnull
}
