% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examples.mdgof.vignette}
\alias{examples.mdgof.vignette}
\title{examples.mdgof.vignette}
\format{
\subsection{'examples.mdgof.vignette'}{

A list
}
}
\usage{
examples.mdgof.vignette
}
\description{
stuff needed to run vignette fast enough to pass CRAN
}
\keyword{datasets}
