% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FF}
\alias{FF}
\title{Find test statistic of Fasano–Franceschini test}
\usage{
FF(dta)
}
\arguments{
\item{dta}{data matrix}
}
\value{
a test statistic
}
\description{
Find test statistic of Fasano–Franceschini test
}
\keyword{internal}
