% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gauss_kernel_matrix}
\alias{gauss_kernel_matrix}
\title{Find gaussian kernel pdf}
\usage{
gauss_kernel_matrix(Eval, S, h)
}
\arguments{
\item{Eval}{a matrix.}

\item{S}{a matrix}

\item{h}{bandwith, a double}
}
\value{
a matrix
}
\description{
Find gaussian kernel pdf
}
