% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.studies.R
\name{gen.cop}
\alias{gen.cop}
\title{Create copula objects}
\usage{
gen.cop(family, p, d = 2)
}
\arguments{
\item{family}{name of copula.}

\item{p}{parameter of copula.}

\item{d}{dimension}
}
\value{
a copula object
}
\description{
This function creates copula objects
}
