% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_eval}
\alias{gen_eval}
\title{Find evaluation points}
\usage{
gen_eval(rnull, p, m)
}
\arguments{
\item{rnull, }{a function that generate new data.}

\item{p, }{a vector of parameters for rnull.}

\item{m, }{size of matrix.}
}
\value{
a matrix
}
\description{
Find evaluation points
}
