% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_mat}
\alias{grad_mat}
\title{Find gradient of log(f) for a matrix of points}
\usage{
grad_mat(x, f)
}
\arguments{
\item{x}{point of evaluation}

\item{f}{function}
}
\value{
a matrix of gradient vectors
}
\description{
Find gradient of log(f) for a matrix of points
}
