% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_vec}
\alias{grad_vec}
\title{Find gradient of log(f)}
\usage{
grad_vec(x, f)
}
\arguments{
\item{x}{point of evaluation}

\item{f}{function}
}
\value{
a gradient vector
}
\description{
Find gradient of log(f)
}
\keyword{internal}
