% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hybrid.mdgof.vignette}
\alias{hybrid.mdgof.vignette}
\title{hybrid.mdgof.vignette}
\format{
\subsection{'hybrid.mdgof.vignette'}{

A list
}
}
\usage{
hybrid.mdgof.vignette
}
\description{
stuff needed to run vignette MDgof::hybrid fast enough to pass CRAN
}
\keyword{datasets}
