% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ksd}
\alias{ksd}
\title{Find test statistic for Kernel Stein Discrepancy test}
\usage{
ksd(X, scf, p)
}
\arguments{
\item{X}{data set.}

\item{scf}{function to find scores}

\item{p}{(possible) parameters}
}
\value{
a double (test statistic)
}
\description{
Find test statistic for Kernel Stein Discrepancy test
}
