% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTSextra.R
\name{makeTSextra}
\alias{makeTSextra}
\title{Create list with needed info}
\usage{
makeTSextra(
  x,
  Continuous,
  pnull,
  rnull,
  phat = function(x) -99,
  dnull = function(x) -99,
  Ranges,
  TSextra
)
}
\arguments{
\item{x}{data set}

\item{Continuous}{=TRUE, is data continuous?}

\item{pnull}{cdf under the null hypothesis}

\item{rnull}{routine to generate data under the null hypothesis}

\item{phat}{=function(x) -99, function to estimate parameters from the data, or -99 if no parameters are estimated}

\item{dnull}{=function(x) -99, density function under the null hypothesis, if available, or -99 if missing}

\item{Ranges}{Range of variables}

\item{TSextra}{(optional) list passed to TS, if needed.}
}
\value{
A list with vectors of test statistics and p.values
}
\description{
This function creates a list with info  needed in various parts of the package
}
