% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdecdf}
\alias{mdecdf}
\title{Finds the empirical distribution function}
\usage{
mdecdf(dta, pts)
}
\arguments{
\item{dta}{a matrix of data points}

\item{pts}{a matrix of evaluation points}
}
\value{
a numeric vector
}
\description{
Finds the empirical distribution function
}
\keyword{internal}
