% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmd}
\alias{mmd}
\title{Find test statistics for continuous data using MMD method}
\usage{
mmd(x, rnull, p)
}
\arguments{
\item{x}{A numeric matrix.}

\item{rnull}{routine to generate data.}

\item{p}{parameter for rnull}
}
\value{
A double
}
\description{
Find test statistics for continuous data using MMD method
}
