% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newTS.R
\name{newTS}
\alias{newTS}
\title{Example for a new test}
\usage{
newTS(x, pnull, p, TSextra)
}
\arguments{
\item{x}{a data set.}

\item{pnull}{function to calculate expected counts.}

\item{p}{parameter for pnull, if needed}

\item{TSextra}{a list with setup info}
}
\value{
a vector with either values of the test statistic, or p values
}
\description{
This shows how a new test routine can be used with Mgof, based on chi square tests
}
