% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{orderC}
\alias{orderC}
\title{R function order(x,y) for Rcpp}
\usage{
orderC(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\value{
a vector of integers
}
\description{
R function order(x,y) for Rcpp
}
\keyword{internal}
