% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p2dC}
\alias{p2dC}
\title{Find probabilities from cdf for discrete data}
\usage{
p2dC(x, cdf, p, Fx = as.numeric(c(-1)))
}
\arguments{
\item{x}{matrix with data}

\item{cdf}{function to find distribution function}

\item{p}{(possible) arguments for cdf}

\item{Fx}{(if available) already calculated values of cdf}
}
\value{
a matrix with probabilities added
}
\description{
Find probabilities from cdf for discrete data
}
