% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_pvals.R
\name{power_pvals}
\alias{power_pvals}
\title{Power estimation of tests that find p values}
\usage{
power_pvals(
  pnull,
  ralt,
  param_alt,
  TS,
  TSextra = list(aa = 0),
  alpha = 0.05,
  B = 1000
)
}
\arguments{
\item{pnull}{cdf function}

\item{ralt}{function that generates data}

\item{param_alt}{parameters for ralt}

\item{TS}{routine that runs the test and returns p values}

\item{TSextra}{=list(aa=0), a list of things passed to TS, if needed}

\item{alpha}{=0.05 type I error probability of test}

\item{B}{=1000 number of simulation runs}
}
\value{
A matrix of power values
}
\description{
This function finds the power for tests that find their own p values
}
\keyword{internal}
