% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power_studies_cont_D5_results}
\alias{power_studies_cont_D5_results}
\title{power_studies_cont_D5_results}
\format{
\subsection{'power_studies_cont_D5_results'}{

A list of matrices with powers
}
}
\usage{
power_studies_cont_D5_results
}
\description{
the results of the included power studies for continuous data without estimation in 5 dimensions
}
\keyword{datasets}
