% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power_studies_cont_nMC5_hybrid_results}
\alias{power_studies_cont_nMC5_hybrid_results}
\title{power_studies_cont_nMC5_hybrid_results}
\format{
\subsection{'power_studies_cont_nMC5_hybrid_results'}{

A list of matrices with powers
}
}
\usage{
power_studies_cont_nMC5_hybrid_results
}
\description{
the results of the included power studies for continuous data without estimation
using two-sample methods and nMC=5
}
\keyword{datasets}
