% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power_studies_disc_results}
\alias{power_studies_disc_results}
\title{power_studies_disc_results}
\format{
\subsection{'power_studies_disc_results'}{

A list of matrices with powers
}
}
\usage{
power_studies_disc_results
}
\description{
the results of the included power studies for discrete data without estimation
}
\keyword{datasets}
