% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RipleyK.R
\name{RipleyK}
\alias{RipleyK}
\title{Ripley's K function test}
\usage{
RipleyK(x)
}
\arguments{
\item{x}{matrix with data}
}
\value{
a number (test statistic)
}
\description{
this function calculates the test statistic of Ripley's K function test
}
