% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rosenblattC}
\alias{rosenblattC}
\title{This function performs a Rosenblatt transform}
\usage{
rosenblattC(x, cdf, p, Range)
}
\arguments{
\item{x}{data set (a matrix)}

\item{cdf}{distribution function}

\item{p}{(possible) parameters for cdf}

\item{Range}{matrix with range of data}
}
\value{
A matrix of transformed data
}
\description{
This function performs a Rosenblatt transform
}
\keyword{internal}
