% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_test_adjusted_pvalue.R
\name{simpvals}
\alias{simpvals}
\title{Helper function to find p values of simulated data.}
\usage{
simpvals(dta, TS, typeTS, TSextra, A, Ranges, nbins, minexpcount, B)
}
\arguments{
\item{dta}{a matrix with data}

\item{TS}{test statistic routine}

\item{typeTS}{type of routine}

\item{TSextra}{a list}

\item{A}{a matrix}

\item{Ranges}{a matrix}

\item{nbins}{a vector}

\item{minexpcount}{an integer}

\item{B}{number of simulation runs}
}
\value{
a matrix
}
\description{
Helper function to find p values of simulated data.
}
\keyword{internal}
