% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_test_adjusted_pvalue.R
\name{simTS}
\alias{simTS}
\title{Helper function to find test statistics of simulated data.}
\usage{
simTS(dta, TS, typeTS, TSextra, B)
}
\arguments{
\item{dta}{a matrix with data}

\item{TS}{test statistic routine}

\item{typeTS}{type of routine}

\item{TSextra}{a list}

\item{B}{number of simulation runs}
}
\value{
a matrix
}
\description{
Helper function to find test statistics of simulated data.
}
\keyword{internal}
