% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.data.R
\name{sq2rec}
\alias{sq2rec}
\title{Rearrange 2D discrete data}
\usage{
sq2rec(x)
}
\arguments{
\item{x}{a matrix of discrete data.}
}
\value{
a rearranged matrix
}
\description{
This function changes a discrete data set given as a nXm counting matrix to a nmX3 matrix
}
