% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{testC}
\alias{testC}
\title{run gof tests for continuous data}
\usage{
testC(dta, rnull, TS, typeTS, TSextra, B = 5000L)
}
\arguments{
\item{dta}{A numeric matrix of data}

\item{rnull}{R function (generate data under null hypothesis)}

\item{TS}{function that calculates test statistics}

\item{typeTS}{integer indicating type of test statistic}

\item{TSextra}{list to pass to TS}

\item{B}{(=5000) Number of simulation runs}
}
\value{
A matrix of numbers (test statistics and p values)
}
\description{
run gof tests for continuous data
}
\keyword{internal}
