% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TS_cont}
\alias{TS_cont}
\title{Find test statistics for continuous data}
\usage{
TS_cont(x, pnull, param, TSextra)
}
\arguments{
\item{x}{A numeric matrix.}

\item{pnull}{cdf.}

\item{param}{parameters for pnull  in case of parameter estimation.}

\item{TSextra}{list with additional info}
}
\value{
A numeric vector with test statistics
}
\description{
Find test statistics for continuous data
}
